Fixes #7, merged branches renamed, and moved classes around
