+++ README.md
@@ -1,1 +1,6 @@
 A wonderfull csv reader
+
+
+Rember to 
+```export PYTHONPATH=. ```
+when running pytest
