+++ setup.py
@@ -1,0 +1,3 @@
+from setuptools import setup, find_packages
+
+setup(name="PACKAGENAME", packages=find_packages())
